requirements.txt
torch
random
numpy
pandas
pillow
scikit-learn
os